added the dummy data into seperate file. it is temp and will be deleted!
