+++ src/data/items.ts
@@ -1,0 +1,29 @@
+  // Dummy data for the items TEMP!!!!
+const items = [
+    {
+      url: "https://t3.ftcdn.net/jpg/07/72/92/66/360_F_772926641_alrxTW34hAXt6Xnnb1mRfD6PrgzwlFLN.jpg",
+      name: "Penguin",
+      category: "Pets",
+      worth: 20,
+    },
+    {
+      url: "https://media.istockphoto.com/id/1008788640/vector/seal-animal-vector-image.jpg?s=612x612&w=0&k=20&c=qZYnxhmtJOxyiiviCtPWlvyJDGlK_HJDBPAXOy41-Gw=",
+      name: "Seal",
+      category: "Pets",
+      worth: 25,
+    },
+    {
+      url: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcStvaFNhjKU6No5CbAa7G6ZOmnbxOh4Ux_ufA&s",
+      name: "Polar Bear",
+      category: "Pets",
+      worth: 30,
+    },
+    {
+      url: "https://t3.ftcdn.net/jpg/06/25/68/32/360_F_625683271_S0JCJuk6uYpGMuKUvY1CneTXAyzzPPzJ.jpg",
+      name: "Fox",
+      category: "Pets",
+      worth: 15,
+    },
+];
+
+export default items
